Let TokenizerTester/TreeTester take directory name
This change allows TokenizerTester and TreeTester to be given a
directory name rather than a list of files (or pathname with a shell
wildcard). And when given a directory name, they recurse the directory
looking for either *.test files (for TokenizerTester) or *.dat files
(for TreeTester), and then run the tests from those files.

Without this change, we can’t easily run TokenizerTester and TreeTester
from AntRun in Maven — because we can’t use shell wildcards in the “arg”
value for the Ant “java” task, and any list of files we otherwise
construct within Maven ends up getting put into the java arg value as a
single string (single argument) — including the spaces between filenames.

So it’s far easier to just give TokenizerTester and TreeTester a
directory name, and use directory recursion within Java to get all the
filenames (as this change does).